requirements.txt
torch>=2.1
biopython>=1.83
fair-esm>=2.0.0
